bot/requirements.txt
git+https://github.com/Rapptz/discord.py
python-dotenv